Rename ssh packages to something more semantic
