+++ service/ssh/key/ssh_addkey.go
@@ -1,10 +1,10 @@
-// +build windows
+// +build darwin linux
 
-package ssh_addkey
+package key
 
 import (
+	"fmt"
 	"github.com/docker/docker/api/types/container"
-	"github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/api/types/network"
 	model "github.com/fubarhouse/pygmy/service/interface"
 )
@@ -12,23 +12,21 @@ import (
 func NewAdder(key string) model.Service {
 	return model.Service{
 		ContainerName: "amazeeio-ssh-agent-add-key",
-		Config:        container.Config{
-			Image:    "amazeeio/ssh-agent",
+		Config: container.Config{
+			Image: "amazeeio/ssh-agent",
 			Cmd: []string{
-				"windows-key-add",
-				"/key",
+				"ssh-add",
+				key,
+			},
+			Labels:		map[string]string{
+				"pygmy": "pygmy",
 			},
 		},
-		HostConfig:    container.HostConfig{
+		HostConfig: container.HostConfig{
+			IpcMode:     "private",
 			AutoRemove:  true,
-			Mounts: []mount.Mount{
-				{
-					Type: "bind",
-					Source: key,
-					Target: key,
-					ReadOnly: true,
-				},
-			},
+			Binds:       []string{fmt.Sprintf("%v:%v", key, key)},
+			VolumesFrom: []string{"amazeeio-ssh-agent"},
 		},
 		NetworkConfig: network.NetworkingConfig{},
 	}
@@ -37,14 +35,17 @@ func NewAdder(key string) model.Service {
 func NewShower() model.Service {
 	return model.Service{
 		ContainerName: "amazeeio-ssh-agent-show-keys",
-		Config:        container.Config{
-			Image:    "amazeeio/ssh-agent",
+		Config: container.Config{
+			Image: "amazeeio/ssh-agent",
 			Cmd: []string{
 				"ssh-add",
 				"-l",
 			},
+			Labels:		map[string]string{
+				"pygmy": "pygmy",
+			},
 		},
-		HostConfig:    container.HostConfig{
+		HostConfig: container.HostConfig{
 			AutoRemove:  true,
 			VolumesFrom: []string{"amazeeio-ssh-agent"},
 		},
