+++ service/ssh/key/ssh_addkey_win.go
@@ -1,10 +1,10 @@
-// +build darwin linux
+// +build windows
 
-package ssh_addkey
+package key
 
 import (
-	"fmt"
 	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/api/types/network"
 	model "github.com/fubarhouse/pygmy/service/interface"
 )
@@ -12,21 +12,23 @@ import (
 func NewAdder(key string) model.Service {
 	return model.Service{
 		ContainerName: "amazeeio-ssh-agent-add-key",
-		Config: container.Config{
-			Image: "amazeeio/ssh-agent",
+		Config:        container.Config{
+			Image:    "amazeeio/ssh-agent",
 			Cmd: []string{
-				"ssh-add",
-				key,
-			},
-			Labels:		map[string]string{
-				"pygmy": "pygmy",
+				"windows-key-add",
+				"/key",
 			},
 		},
-		HostConfig: container.HostConfig{
-			IpcMode:     "private",
+		HostConfig:    container.HostConfig{
 			AutoRemove:  true,
-			Binds:       []string{fmt.Sprintf("%v:%v", key, key)},
-			VolumesFrom: []string{"amazeeio-ssh-agent"},
+			Mounts: []mount.Mount{
+				{
+					Type: "bind",
+					Source: key,
+					Target: key,
+					ReadOnly: true,
+				},
+			},
 		},
 		NetworkConfig: network.NetworkingConfig{},
 	}
@@ -35,17 +37,14 @@ func NewAdder(key string) model.Service {
 func NewShower() model.Service {
 	return model.Service{
 		ContainerName: "amazeeio-ssh-agent-show-keys",
-		Config: container.Config{
-			Image: "amazeeio/ssh-agent",
+		Config:        container.Config{
+			Image:    "amazeeio/ssh-agent",
 			Cmd: []string{
 				"ssh-add",
 				"-l",
 			},
-			Labels:		map[string]string{
-				"pygmy": "pygmy",
-			},
 		},
-		HostConfig: container.HostConfig{
+		HostConfig:    container.HostConfig{
 			AutoRemove:  true,
 			VolumesFrom: []string{"amazeeio-ssh-agent"},
 		},
